some test fixes
