+++ internal/controller/controller_test.go
@@ -1,14 +1,44 @@
 package controller
 
 import (
+	"github.com/asecurityteam/rolling"
 	"github.com/markusressel/fan2go/internal"
 	"github.com/markusressel/fan2go/internal/configuration"
 	"github.com/markusressel/fan2go/internal/curves"
 	"github.com/markusressel/fan2go/internal/fans"
-	"github.com/markusressel/fan2go/internal/testingutils"
 	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
+)
+
+var (
+	LinearFan = map[int][]float64{
+		0:   {0.0},
+		255: {255.0},
+	}
+
+	NeverStoppingFan = map[int][]float64{
+		0:   {50.0},
+		50:  {50.0},
+		255: {255.0},
+	}
+
+	CappedFan = map[int][]float64{
+		0:   {0.0},
+		1:   {0.0},
+		2:   {0.0},
+		3:   {0.0},
+		4:   {0.0},
+		5:   {0.0},
+		6:   {20.0},
+		200: {200.0},
+	}
+
+	CappedNeverStoppingFan = map[int][]float64{
+		0:   {50.0},
+		50:  {50.0},
+		200: {200.0},
+	}
 )
 
 type mockPersistence struct{}
@@ -19,12 +49,36 @@ func (p mockPersistence) LoadFanPwmData(fan fans.Fan) (map[int][]float64, error)
 	return fanCurveDataMap, nil
 }
 
+func CreateFan(neverStop bool, curveData map[int][]float64) (fan fans.Fan, err error) {
+	configuration.CurrentConfig.RpmRollingWindowSize = 10
+
+	fan = &fans.HwMonFan{
+		Config: configuration.FanConfig{
+			ID: "fan1",
+			HwMon: &configuration.HwMonFanConfig{
+				Platform: "platform",
+				Index:    1,
+			},
+			NeverStop: neverStop,
+			Curve:     "curve",
+		},
+		FanCurveData: &map[int]*rolling.PointPolicy{},
+		PwmOutput:    "fan1_output",
+		RpmInput:     "fan1_rpm",
+	}
+	fans.FanMap[fan.GetConfig().ID] = fan
+
+	err = fan.AttachFanCurveData(&curveData)
+
+	return fan, err
+}
+
 func TestLinearFan(t *testing.T) {
 	// GIVEN
-	fan, _ := testingutils.CreateFan(false, testingutils.LinearFan)
+	fan, _ := CreateFan(false, LinearFan)
 
 	// WHEN
-	startPwm, maxPwm := ComputePwmBoundaries(fan)
+	startPwm, maxPwm := fans.ComputePwmBoundaries(fan)
 
 	// THEN
 	assert.Equal(t, 1, startPwm)
@@ -33,10 +87,10 @@ func TestLinearFan(t *testing.T) {
 
 func TestNeverStoppingFan(t *testing.T) {
 	// GIVEN
-	fan, _ := testingutils.CreateFan(false, testingutils.NeverStoppingFan)
+	fan, _ := CreateFan(false, NeverStoppingFan)
 
 	// WHEN
-	startPwm, maxPwm := ComputePwmBoundaries(fan)
+	startPwm, maxPwm := fans.ComputePwmBoundaries(fan)
 
 	// THEN
 	assert.Equal(t, 0, startPwm)
@@ -45,10 +99,10 @@ func TestNeverStoppingFan(t *testing.T) {
 
 func TestCappedFan(t *testing.T) {
 	// GIVEN
-	fan, _ := testingutils.CreateFan(false, testingutils.CappedFan)
+	fan, _ := testingutils.CreateFan(false, CappedFan)
 
 	// WHEN
-	startPwm, maxPwm := ComputePwmBoundaries(fan)
+	startPwm, maxPwm := fans.ComputePwmBoundaries(fan)
 
 	// THEN
 	assert.Equal(t, 6, startPwm)
@@ -57,10 +111,10 @@ func TestCappedFan(t *testing.T) {
 
 func TestCappedNeverStoppingFan(t *testing.T) {
 	// GIVEN
-	fan, _ := testingutils.CreateFan(false, testingutils.CappedNeverStoppingFan)
+	fan, _ := CreateFan(false, CappedNeverStoppingFan)
 
 	// WHEN
-	startPwm, maxPwm := ComputePwmBoundaries(fan)
+	startPwm, maxPwm := fans.ComputePwmBoundaries(fan)
 
 	// THEN
 	assert.Equal(t, 0, startPwm)
@@ -70,7 +124,7 @@ func TestCappedNeverStoppingFan(t *testing.T) {
 func TestCalculateTargetSpeedLinear(t *testing.T) {
 	// GIVEN
 	avgTmp := 50000.0
-	s := testingutils.CreateSensor(
+	s := CreateSensor(
 		"sensor",
 		configuration.HwMonSensorConfig{
 			Platform: "platform",
@@ -87,7 +141,7 @@ func TestCalculateTargetSpeedLinear(t *testing.T) {
 	)
 	curve, _ := curves.NewSpeedCurve(curveConfig)
 
-	fan, _ := internal.createFan(false, linearFan)
+	fan, _ := CreateFan(false, linearFan)
 
 	controller := fanController{
 		mockPersistence{},
@@ -126,7 +180,7 @@ func TestCalculateTargetSpeedNeverStop(t *testing.T) {
 	)
 	curve, _ := curves.NewSpeedCurve(curveConfig)
 
-	fan, _ := internal.createFan(true, cappedFan)
+	fan, _ := CreateFan(true, cappedFan)
 
 	controller := fanController{
 		mockPersistence{},
@@ -140,7 +194,7 @@ func TestCalculateTargetSpeedNeverStop(t *testing.T) {
 	if err != nil {
 		assert.Fail(t, err.Error())
 	}
-	target := controller.calculateTargetPwm(fan, 0, optimal)
+	target := calculateTargetPwm(fan, 0, optimal)
 
 	// THEN
 	assert.Equal(t, 0, optimal)
