+++ internal/fans/file.go
@@ -70,6 +70,10 @@ func (fan *FileFan) SetFanCurveData(data *map[int]*rolling.PointPolicy) {
 	panic("implement me")
 }
 
+func (fan *FileFan) AttachFanCurveData(curveData *map[int][]float64) (err error) {
+	panic("implement me")
+}
+
 func (fan FileFan) GetPwmEnabled() (int, error) {
 	panic("implement me")
 }
