+++ internal/persistence/persistence_test.go
@@ -3,7 +3,6 @@ package persistence
 import (
 	"github.com/asecurityteam/rolling"
 	"github.com/markusressel/fan2go/internal/configuration"
-	"github.com/markusressel/fan2go/internal/controller"
 	"github.com/markusressel/fan2go/internal/fans"
 	"github.com/stretchr/testify/assert"
 	"testing"
@@ -13,11 +12,24 @@ const (
 	dbTestingPath = "./test.db"
 )
 
+var (
+	LinearFan = map[int][]float64{
+		0:   {0.0},
+		255: {255.0},
+	}
+
+	NeverStoppingFan = map[int][]float64{
+		0:   {50.0},
+		50:  {50.0},
+		255: {255.0},
+	}
+)
+
 func TestWriteFan(t *testing.T) {
 	// GIVEN
 	p := NewPersistence(dbTestingPath)
 
-	fan, _ := createFan(false, linearFan)
+	fan, _ := createFan(false, LinearFan)
 
 	// WHEN
 	err := p.SaveFanPwmData(fan)
@@ -30,10 +42,10 @@ func TestReadFan(t *testing.T) {
 	// GIVEN
 	persistence := NewPersistence(dbTestingPath)
 
-	fan, _ := createFan(false, neverStoppingFan)
+	fan, _ := createFan(false, NeverStoppingFan)
 	err := persistence.SaveFanPwmData(fan)
 
-	fan, _ = createFan(false, linearFan)
+	fan, _ = createFan(false, LinearFan)
 
 	// WHEN
 	fanData, err := persistence.LoadFanPwmData(fan)
@@ -62,7 +74,7 @@ func createFan(neverStop bool, curveData map[int][]float64) (fan fans.Fan, err e
 	}
 	fans.FanMap[fan.GetConfig().ID] = fan
 
-	err = controller.AttachFanCurveData(&curveData, fan)
+	err = fan.AttachFanCurveData(&curveData)
 
 	return fan, err
 }
